Added content to get-started
added the content for cloud, linux, and macOS(M1) to general.md

MacOS content needs to be modified more, possibly explain that brew
can be used instead of just manually downloading the content. Do the
same with the linux section maybe?

TODO: set up WSL on windows and test if local install works.