Update to changelogs data:
--- changes ---

content/connectors/*/*.md: added a changelog shortcode call. This is
both to prove that it works for all the documents, and to just have
all the files have the changelog data. That was
content/connectors/source/http.md doesn't feel lonely.

layouts/shortcodes/changelog.html: Added a couple comments to the
script, to specify how the format should be (don't need to have the
leading `content/connectors` nor the `CHANGELOG.md` on the changelog
call). Added checking to see if the file existed, and if not, to
produce the default empty changelog file.

changelogs/compiled/default-changelog.md: The default empty changelog
produced by the shortcodes if the requested changelog does not exist.

changelogs/readme.md: a readme for the changelog information, to tell
users how to use it manually. Basically do not touch anything in the
/changelogs/compiled directory.

scripts/changelog.py: made len(".changelogs/compiled") and the magic
number 18 into a global bariable `LENGTH_OF_START_PATH`.

--- comments ---

Will need to investigate setting up changelog.yml to execute from the
fluvio-connectors repository instead.